+++ src/main/java/com/esprit/tn/services/ISubscriptionService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.esprit.tn.entities.Subscription;
 
 public interface ISubscriptionService {
-	public void ActivateSubscription(Subscription sub, int userid);
+	public void ActivateSubscription(Subscription sub, long userid);
 	public List<Subscription> getAllSubscriptions();
 	public Subscription getSubscriptionById(int id);
 	public Subscription updateSubscription(Subscription s , int id);
